requirements.txt
discord.py
requests
python-dotenv
google-genai
flask
